move more functions out of the buffer module
